Import non-WASM-related modules via path aliases
For the time signal generator module, using an alias fails
when bundled because we're treating the module and its JS
glue as Rollup externals. This is necessary because:

  - a .wasm binary cannot be loaded via `new URL()` in
    AudioWorkletGlobalScope, so the loader fails init due
    to the missing symbol, and Rollup didn't realize it
    should have bundled timesignal.wasm anyway
  - the Audio Worklet (timesignal.aw.js) is being loaded
    from an actual .js file, not a Blob URL (for now)

The compute-only edit distance WASM module has no such
issues and could be imported via a path alias, but we
wouldn't want the time signal generator to get lonely.

Also, reorganize imports while we're at it.