Initial Progressive Web App
Uses the Vite PWA plugin and workbox. workbox becomes a
runtime dependency, but it's only a conceptual difference.

The Service Worker is module-based, precaching all site
assets and adding COOP/COEP headers if needed.

Note that the Vite PWA plugin generates a non-module-based
Service Worker, but only in prod. Firefox doesn't support
module-based Service Workers yet. So we shall live with
disabling the Service Worker (and the PWA functionality) in
dev and adding the headers in the Vite server configuration.
This allows manual testing with Firefox.

While we're at it, do some UI stuff and remove the Vite
basic SSL plugin. A self-signed certificate does not result
in a secure enough context for a Service Worker without
adding it as a root certificate, and that's annoying, so
the plugin no longer serves its intended purpose of
allowing manual testing from the same LAN.